Refactor: Move struct Bitmap into image.hh
