+++ game/surface.cc
@@ -2,7 +2,6 @@
 
 #include "configuration.hh"
 #include "video_driver.hh"
-#include "image.hh"
 #include "screen.hh"
 #include "svg.hh"
 #include <boost/algorithm/string/case_conv.hpp>
